1.Modified package and renamed function name getBucketName to getFileNameFromDelim. 2.Added reference link on method.
