+++ api-testing/multicloud-api/src/main/java/com/soda/utils/Constant.java
@@ -1,4 +1,4 @@
-package com.opensds.utils;
+package com.soda.utils;
 
 /**
  * In this class set Inputs path.
