+++ api-testing/multicloud-api/src/main/java/com/soda/utils/HeadersName.java
@@ -1,4 +1,4 @@
-package com.opensds.utils;
+package com.soda.utils;
 
 /**
  * This class use for HTTP Headers Name.
